% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{cancel_order}
\alias{cancel_order}
\title{Function to cancel an exiting order.}
\usage{
cancel_order(object, orderid, variety)
}
\arguments{
\item{object}{An object of type Smart connect.}

\item{orderid}{Order ID for the order to modify.}

\item{variety}{Order variety, e.g. regular, bo, co or amo.}
}
\value{
Returns an order ID (string), if successful.
}
\description{
Function to cancel an exiting order with specified ID.
}
\details{
This places a request to cancel an existing order, if it is not
already executed. Please note: no error checks are done for the input
parameters. Sanity check is left to the user of this function. If
successful, this will return the order ID (should be same as the
one passed). A successfully placed request does not mean it is
a valid order, and orders with erroneous user input can be immediately
canceled. Therefore it is good practice to check the order details once
an order ID is received from this call.
}
