% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{generate_session}
\alias{generate_session}
\title{Function to start an api session.}
\usage{
generate_session(object, clientCode, password)
}
\arguments{
\item{object}{of type smart connect, must have a valid client code and password set.}

\item{clientCode}{valid client id of the user is must.}

\item{password}{valid password against the user/client id is must.}
}
\value{
the object data with access token and refresh token.
}
\description{
This function is used to login the session.
}
\details{
This function sends a request to smart api to login the session.
}
