% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{generate_token}
\alias{generate_token}
\title{Fetch access token given the refresh token.}
\usage{
generate_token(object,refresh_token)
}
\arguments{
\item{object}{An object of type smart connect, must have refresh token set.}

\item{refresh_token}{refresh token is passed.}
}
\value{
A string containing the access token.
}
\description{
This function returns an access token given the inputs, that
can be used for subsequent api calls.
}
\details{
This function generate the `access_token` by exchanging
`refresh_token`.The `refresh_token` obtained after the login flow. the
`access_token` required for all subsequent requests. The object passed in
this call must already have `refresh_token` set. A successful call also set the user data within the object.
}
