% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{get_candle_data}
\alias{get_candle_data}
\title{Function to get the historic data}
\usage{
get_candle_data(object, exchange, symboltoken, interval, fromdate, todate)
}
\arguments{
\item{object}{An object of type smart connect.}

\item{exchange}{Exchange for the order, e.g. NSE, NFO, CDS etc}

\item{symboltoken}{symboltoken.}

\item{interval}{time interval}

\item{fromdate}{fromdate}

\item{todate}{todate}
}
\value{
Returns historic data from the date range given with the time interval provided, if successful.
}
\description{
Function to get the historic data
}
