% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{get_feed_token}
\alias{get_feed_token}
\title{Function to get the feed_token.}
\usage{
get_feed_token(object)
}
\arguments{
\item{object}{An object of smart connect type.}
}
\value{
Returns the feed_token.
}
\description{
Function to get the feed_token from
the smart object.
}
