% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{get_ltp_data}
\alias{get_ltp_data}
\title{Function to obtain last traded prices.}
\usage{
get_ltp_data(object, exchange, tradingsymbol, symboltoken)
}
\arguments{
\item{object}{An object of type smart connect.}

\item{exchange}{Exchange for the order, e.g. NSE, NFO, CDS etc.}

\item{tradingsymbol}{Trading Symbol of the instrument.}

\item{symboltoken}{Symbol token}
}
\value{
a data frame with last price data, with instruments as row names.
}
\description{
Function to obtain last traded prices.
}
\details{
This function accepts a list of instruments (as exchange:symbol
pair, e.g. NFO:NIFTY18JANFUT) and returns last traded prices. Please note:
this function is meant to be called as a one-off way to get the quotes.
One should not call this function too many times. The api possibly has a
rate limit and too many requests will block the user
}
