% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{get_profile}
\alias{get_profile}
\title{Function to get the details of logged on user.}
\usage{
get_profile(object,refresh_token)
}
\arguments{
\item{object}{An object of type smart connect with valid refresh_token.}

\item{refresh_token}{refresh_token needs to be passed}
}
\value{
Returns a list with user details, if successful.
}
\description{
Gets the details of the user.
}
