% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{gtt_cancel}
\alias{gtt_cancel}
\title{Function to cancel gtt rule}
\usage{
gtt_cancel(object, id, symboltoken, exchange)
}
\arguments{
\item{object}{An object of type smart connect.}

\item{id}{Gtt rule id.}

\item{symboltoken}{Symbol token}

\item{exchange}{Exchange for the order, e.g. NSE, NFO, CDS etc}
}
\value{
Returns a rule id, if successful.
}
\description{
Function to cancel gtt rule
}
