% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{gtt_create}
\alias{gtt_create}
\title{Function to create Gtt rules}
\usage{
gtt_create(
  object,
  tradingsymbol,
  symboltoken,
  exchange,
  producttype,
  transactiontype,
  price,
  quantity,
  disclosedqty,
  triggerprice,
  timeperiod
)
}
\arguments{
\item{object}{An object of type smart connect}

\item{tradingsymbol}{Trading Symbol of the instrument.}

\item{symboltoken}{Symbol token}

\item{exchange}{Exchange for the order, e.g. NSE, NFO, CDS etc.}

\item{producttype}{Product type, e.g. MIS, NRML, CNC, BO or CO.}

\item{transactiontype}{BUY or SELL.}

\item{price}{New Price (e.g. for limit orders).}

\item{quantity}{New quantity.}

\item{disclosedqty}{Disclosed Quantity}

\item{triggerprice}{Trigger Price}

\item{timeperiod}{Time Period.}
}
\value{
Returns a rule id, if successful.
}
\description{
Function to create Gtt rules
}
