% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{gtt_details}
\alias{gtt_details}
\title{Function to get the details of gtt rule}
\usage{
gtt_details(object, id)
}
\arguments{
\item{object}{An object of type smart connect.}

\item{id}{Gtt rule id.}
}
\value{
Returns a data frame with gtt details, if successful.
}
\description{
Function to get the details of gtt rule
}
