% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{gtt_lists}
\alias{gtt_lists}
\title{Function to get the detailed list of gtt rule}
\usage{
gtt_lists(object, status, page, count)
}
\arguments{
\item{object}{An object of type smart connect.}

\item{status}{a list of status to find the rulelist accordingly}

\item{page}{No of page}

\item{count}{no of counts of rule}
}
\value{
Returns a data frame with gtt detailed list, if successful.
}
\description{
Function to get the detailed list of gtt rule
}
