% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{gtt_modify}
\alias{gtt_modify}
\title{Function to modify gtt rule}
\usage{
gtt_modify(
  object,
  id,
  tradingsymbol,
  symboltoken,
  exchange,
  producttype,
  transactiontype,
  price,
  quantity,
  triggerprice,
  disclosedqty,
  timeperiod
)
}
\arguments{
\item{object}{An object of type smart connect}

\item{id}{Gtt rule id}

\item{symboltoken}{Symbol token}

\item{exchange}{Exchange for the order, e.g. NSE, NFO, CDS etc.}

\item{price}{New Price (e.g. for limit orders).}

\item{quantity}{New quantity.}

\item{triggerprice}{Trigger Price}

\item{disclosedqty}{Disclosed Quantity}

\item{timeperiod}{Time Period.}
}
\value{
Returns a rule id, if successful.
}
\description{
Function to modify gtt rule
}
