% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{holding}
\alias{holding}
\title{Function to get the holding details.}
\usage{
holding(object)
}
\arguments{
\item{object}{An object of type smart connect.}
}
\value{
Returns a list with holdings details, if successful.
}
\description{
Gets the holdings details of the user.
}
