% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{modify_order}
\alias{modify_order}
\title{Function to modify an exiting order.}
\usage{
modify_order(object, param)
}
\arguments{
\item{object}{An object of type smart connect.}

\item{variety}{Order variety, e.g. regular, bo, co or amo.}

\item{orderid}{Order ID for the order to modify.}

\item{ordertype}{Order type, e.g. MARKET, LIMIT, SL or SL-M.}

\item{duration}{Duration}

\item{producttype}{Product type, e.g. MIS, NRML, CNC, BO or CO}

\item{price}{New Price (e.g. for limit orders).}

\item{quantity}{New quantity.}

\item{tradingsymbol}{Trading Symbol of the instrument.}

\item{symboltoken}{Symbol token}

\item{exchange}{Exchange for the order, e.g. NSE, NFO, CDS etc.}
}
\value{
Returns an order ID (string), if successful.
}
\description{
Function to modify an exiting order with specified details.
}
\details{
This function sends a modification request for an existing order,
assuming it is not already executed. Please note: not all parameters are
relevant for all types of orders. No error checks is carried out, It is
left to the function calling this method to carry out necessary error
checks.If successful, this will return the order ID (should be same as the
one passed). A successfully placed request does not mean it is
a valid order, and orders with erroneous user input can be immediately
canceled. Therefore it is good practice to check the order details once
an order ID is received from this call.
}
