% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{order_book}
\alias{order_book}
\title{Function to get the order details.}
\usage{
order_book(object)
}
\arguments{
\item{object}{An object of type smart connect.}
}
\value{
Returns a data frame with order details, if successful.
}
\description{
Gets the order details of the user.
}
