% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{place_order}
\alias{place_order}
\title{Function to place a trade.}
\usage{
place_order(object, param)
}
\arguments{
\item{object}{An object of type smart connect.}

\item{variety}{Order variety, e.g. regular, bo, co or amo.}

\item{tradingsymbol}{Trading Symbol of the instrument.}

\item{symboltoken}{Symbol token}

\item{transactiontype}{BUY or SELL.}

\item{exchange}{Exchange for the order, e.g. NSE, NFO, CDS etc.}

\item{ordertype}{Order type, e.g. MARKET, LIMIT, SL or SL-M.}

\item{producttype}{Product type, e.g. MIS, NRML, CNC, BO or CO.}

\item{duration}{Duration}

\item{price}{Price if required (e.g. for limit orders).}

\item{squareoff}{Price difference for profit booking (bracker orders).}

\item{stoploss}{Price difference for loss booking (bracket orders).}

\item{quantity}{Quantity of the order.}
}
\value{
Returns an order ID (string), if successful.
}
\description{
Function to place a trade with specified details.
}
\details{
This function sends an order with input details for execution.
Please note not all parameters are relevant for all types of orders. This
function (as with rest of the package) does no error checks and post the
trade request as is. It is left to the function calling this method to
carry out necessary error checks. If the order is successfully posted, an
order ID will be returned. A successfully posted order does not mean it is
a valid order, and orders with erroneous user input can be immediately
cancelled. Therefore it is good practice to check the order details once
an order ID is received from this call.
}
