% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{position}
\alias{position}
\title{Function to get the position details.}
\usage{
position(object)
}
\arguments{
\item{object}{An object of type smart connect.}
}
\value{
Returns a dataframe with trade details, if successful.
}
\description{
Gets the position details of the user.
}
