% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{rms_limit}
\alias{rms_limit}
\title{Function to get the rms limit details.}
\usage{
rms_limit(object)
}
\arguments{
\item{object}{An object of type smart connect.}
}
\value{
Returns with rms_limit details, if successful.
}
\description{
Gets the rms limit details of the user.
}
