% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{set_access_token}
\alias{set_access_token}
\title{Function to set the access token in the smart connect object.}
\usage{
set_access_token(object, access_token)
}
\arguments{
\item{object}{a smart connect object.}

\item{access_token}{Access token.}
}
\value{
Smart connect object with access token set.
}
\description{
This function sets the access token inside the smart connect object.
}
