% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{set_feed_token}
\alias{set_feed_token}
\title{Function to set the feed token in the smart connect object.}
\usage{
set_feed_token(object, feed_token)
}
\arguments{
\item{object}{a smart connect object.}

\item{feed_token}{Feed token.}
}
\value{
Smart connect object with feed token set.
}
\description{
This function sets the feed token inside the smart connect object.
}
