% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{set_refresh_token}
\alias{set_refresh_token}
\title{Function to set the refresh token in the smart connect object.}
\usage{
set_refresh_token(object, refresh_token)
}
\arguments{
\item{object}{an S4 object of type smart connect.}

\item{refresh_token}{a string containing the refresh token from login flow.}
}
\value{
Returns the object with refresh token field set.
}
\description{
Function to set refresh token (obtained from a successful login flow) in the smart connect object.
}
\details{
This function allows to set the refresh token of the smart
connect object. The refresh token is obtained from a successful login flow.
After the login flow, the user is redirected to a pre-specified URL. The
refresh token is posted alongwith as a query parameter.
}
