% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{set_user_id}
\alias{set_user_id}
\title{Function to set the user_id in the smart connect object.}
\usage{
set_user_id(object, user_id)
}
\arguments{
\item{object}{a smart connect object.}

\item{user_id}{Access token.}
}
\value{
Smart connect object with user_id set.
}
\description{
This function sets the user_id inside the smart connect object.
}
