% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{terminate_session}
\alias{terminate_session}
\title{Function to end an api session.}
\usage{
terminate_session(object, user_id)
}
\arguments{
\item{object}{of type smart connect, must have a valid client code.}

\item{user_id}{it is the client id of the user}
}
\value{
An object of type smart connect.
}
\description{
This function is used to logout the session.
}
\details{
This function sends a request to smart api to logout the session.
}
