% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{trade_book}
\alias{trade_book}
\title{Function to get the trade details.}
\usage{
trade_book(object)
}
\arguments{
\item{object}{An object of type smart connect.}
}
\value{
Returns a data frame with trade details, if successful.
}
\description{
Gets the order details of the user.
}
